% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{dataShpReg}
\alias{dataShpReg}
\title{dataShpReg}
\usage{
dataShpReg(chemin)
}
\arguments{
\item{chemin}{valeur de chemin : l'eplacement ou ce trouve le fichier shapefile}
}
\value{
data shapefile
}
\description{
Cette fonction permet de récupérer le contenu (data) du fichier shapefile de regions du Maroc
}
\examples{
dataShpReg("data/shp_nouveau_decoupage_territorial_2015")
}
