% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{mapCsvReg}
\alias{mapCsvReg}
\title{mapCsvReg}
\usage{
mapCsvReg(num, col, we)
}
\arguments{
\item{num}{de la region a afficher}

\item{col}{valeur de la couleur des frontieres des regions}

\item{we}{valeur de l'paisseur es frontieres des regions}
}
\description{
Cette fonction permet de tracer une region choisis du Maroc
}
\examples{
output$regions <- renderLeaflet({
mapCsvReg(6,"red",1.25)
})
}
