% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{mapCsvRegs}
\alias{mapCsvRegs}
\title{mapCsvRegs}
\usage{
mapCsvRegs(col, we)
}
\arguments{
\item{col}{valeur de la couleur des frontieres des regions}

\item{we}{valeur de l'paisseur es frontieres des regions}
}
\description{
Cette fonction permet de tracer la carte du Maroc - regions
}
\examples{
output$regions <- renderLeaflet({
mapCsvRegs("red",1.25)
})
}
