% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{mapShpDensPop}
\alias{mapShpDensPop}
\title{mapShpDensPop}
\usage{
mapShpDensPop(shp, reg, coul, dens)
}
\arguments{
\item{shp}{valeur de data du fichier shapefile}

\item{reg}{valeur de la region a afficher}

\item{coul}{valeur de la couleur de la region choisis}
}
\description{
Cette fonction permet de selectionner la region dont son nom passe au parametre
}
\examples{
shp <- dataShpReg("data/shp_nouveau_decoupage_territorial_2015")
reg = "Fes-Meknes"
mapShpDensPop(shp,reg,"green")
}
