% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{plotSHP}
\alias{plotSHP}
\title{plotSHP}
\usage{
plotSHP(chemin)
}
\arguments{
\item{chemin}{valeur de chemin : l'eplacement ou ce trouve le fichier shapefile}
}
\description{
Cette fonction permet de ploter la carte des regions du Maroc
}
\examples{
plotSHP("data/shp_nouveau_decoupage_territorial_2015")
}
