% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{uploadSHP}
\alias{uploadSHP}
\title{uploadSHP}
\usage{
uploadSHP()
}
\arguments{
\item{sans}{parametres}
}
\description{
Cette fonction permet de charger les fichiers  shapefile, ou « fichier de formes » est un format de fichier pour les systèmes d'informations géographiques (SIG) pour les régions du Maroc
}
\examples{
uploadSHP()
}
